10.3

+++ auth-main/PhotosService/Controllers/PhotosApiController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
@@ -25,8 +26,10 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetPhotos(string ownerId)
+        public async Task<IActionResult> GetPhotos(string ownerId, JwtSecurityToken accessToken)
         {
+            if (accessToken.Subject != ownerId)
+                return Forbid();
             var photoEntities = (await photosRepository.GetPhotosAsync(ownerId)).ToList();
             var photos = mapper.Map<List<PhotoDto>>(photoEntities);
             foreach(var photo in photos)
@@ -35,20 +38,26 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet("{id}/meta")]
-        public async Task<IActionResult> GetPhotoMeta(Guid id)
+        public async Task<IActionResult> GetPhotoMeta(Guid id, JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
-
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
             var photo = mapper.Map<PhotoDto>(photoEntity);
             photo.Url = GeneratePhotoUrl(photo);
             return Ok(photo);
         }
 
         [HttpGet("{id}/content")]
-        public async Task<IActionResult> GetPhotoContent(Guid id)
+        public async Task<IActionResult> GetPhotoContent(Guid id, JwtSecurityToken accessToken)
         {
+            var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
+            if (photoEntity == null)
+                return NotFound();
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
             var photoContent = await photosRepository.GetPhotoContentAsync(id);
             if (photoContent == null)
                 return NotFound();
@@ -57,8 +66,10 @@ namespace PhotosService.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> AddPhoto(PhotoToAddDto photo)
+        public async Task<IActionResult> AddPhoto(PhotoToAddDto photo, [FromHeader] JwtSecurityToken accessToken)
         {
+            if (accessToken.Subject != photo.OwnerId)
+                return Forbid();
             var content = Convert.FromBase64String(photo.Base64Content);
             var result = await photosRepository.AddPhotoAsync(photo.Title, photo.OwnerId, content);
             if (!result)
@@ -67,12 +78,13 @@ namespace PhotosService.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdatePhoto(Guid id, PhotoToUpdateDto photo)
+        public async Task<IActionResult> UpdatePhoto(Guid id, PhotoToUpdateDto photo, [FromHeader]  JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
-
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
             photoEntity.Title = photo.Title;
             var result = await photosRepository.UpdatePhotoAsync(photoEntity);
             if (!result)
@@ -81,12 +93,13 @@ namespace PhotosService.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeletePhoto(Guid id)
+        public async Task<IActionResult> DeletePhoto(Guid id, JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
-
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
             var result = await photosRepository.DeletePhotoAsync(photoEntity);
             if (!result)
                 return Conflict();
